Add ping and pingtest modules.
The ping module is a wrapper around the OS's ping command, allowing
Python to ping other machines without needing to be setuid as root.

The pingtest module uses the ping module to generate ping statistics for
one or more hosts.

Eventually, this can be used to show a dashboard of the latest ping
times to one or more hosts.